requirements.txt
Django>=5.1.7
psycopg2-binary>=2.9.6
django-environ>=0.10.0
gunicorn>=21.2.0
whitenoise>=6.4.0
Pillow>=9.5.0
celery>=5.4.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0